update lending pool adapter environment var handling logic
