+++ huma_signals/commons/chains.py
@@ -1,11 +1,6 @@
 from __future__ import annotations
 
 import enum
-import os
-from typing import Any, Dict, Optional, Sequence, Type, Union
-
-import web3
-from web3 import eth, geth, middleware, module, net
 
 
 class Chain(enum.Enum):
@@ -28,65 +23,3 @@ class Chain(enum.Enum):
 
     def is_testnet(self) -> bool:
         return self.chain_name() in ("goerli")
-
-
-def get_w3(chain: Chain, alchemy_key: Optional[str] = None) -> web3.Web3:
-    """Get a web3 instance for the given chain
-
-    Parameters
-    ----------
-    chain : Chain
-        the chain to connect to
-    alchemy_key : str, optional
-        the alchemy_key to connect to the chain with, by default None
-        when set to None, function will try to get from env
-
-    Returns
-    -------
-    Web3
-        the web3 instance
-    """
-    if not alchemy_key:
-        # Try to get from env
-        alchemy_key = os.getenv(f"ALCHEMY_KEY_{chain.name.upper()}")
-
-    if not alchemy_key:
-        raise ValueError(f"Alchemy key not set for chain: {chain}")
-
-    modules: Dict[str, Union[Type[module.Module], Sequence[Any]]] = {
-        "eth": eth.AsyncEth,
-        "net": net.AsyncNet,
-        "geth": (
-            geth.Geth,
-            {
-                "txpool": geth.AsyncGethTxPool,
-                "personal": geth.AsyncGethPersonal,
-                "admin": geth.AsyncGethAdmin,
-            },
-        ),
-    }
-    if chain == Chain.ETHEREUM:
-        return web3.Web3(
-            provider=web3.Web3.AsyncHTTPProvider(
-                f"https://eth-mainnet.g.alchemy.com/v2/{alchemy_key}"
-            ),
-            modules=modules,
-        )
-    if chain == Chain.POLYGON:
-        return web3.Web3(
-            provider=web3.Web3.AsyncHTTPProvider(
-                f"https://polygon-mainnet.g.alchemy.com/v2/{alchemy_key}"
-            ),
-            modules=modules,
-        )
-    if chain == Chain.GOERLI:
-        w3 = web3.Web3(
-            provider=web3.Web3.AsyncHTTPProvider(
-                f"https://eth-goerli.g.alchemy.com/v2/{alchemy_key}"
-            ),
-            modules=modules,
-        )
-        w3.middleware_onion.inject(middleware.async_geth_poa_middleware, layer=0)
-        return w3
-
-    raise ValueError(f"Unsupported chain: {chain}")
